+++ DuckDuckGo/OnboardingExperiment/OnboardingIntro/OnboardingView+AddressBarPositionContent.swift
@@ -34,8 +34,6 @@ extension OnboardingView {
     }
 
     struct AddressBarPositionContent: View {
-        
-        @StateObject private var viewModel = AddressBarPositionContentViewModel()
 
         private var animateTitle: Binding<Bool>
         private var showContent: Binding<Bool>
@@ -60,7 +58,7 @@ extension OnboardingView {
                 .font(Metrics.titleFont)
 
                 VStack(spacing: 24) {
-                    addressBarPositionButtons
+                    OnboardingAddressBarPositionPicker()
 
                     Button(action: action) {
                         Text(verbatim: UserText.HighlightsOnboardingExperiment.AddressBarPosition.cta)
@@ -69,97 +67,6 @@ extension OnboardingView {
                 }
                 .visibility(showContent.wrappedValue ? .visible : .invisible)
             }
-        }
-
-        private var addressBarPositionButtons: some View {
-            VStack(spacing: 16.0) {
-                ForEach(viewModel.items, id: \.type) { item in
-                    AddressBarPositionButton(
-                        icon: item.icon,
-                        title: AttributedString(item.title),
-                        message: item.message,
-                        isSelected: item.isSelected,
-                        action: {
-                            viewModel.setAddressBar(position: item.type)
-                        }
-                    )
-                }
-            }
-        }
-    }
-
-}
-
-// MARK: - ViewModel
-
-final class AddressBarPositionContentViewModel: ObservableObject {
-
-    struct AddressBarPositionDisplayModel {
-        let type: AddressBarPosition
-        let icon: ImageResource
-        let title: NSAttributedString
-        let message: String
-        let isSelected: Bool
-    }
-
-    @Published private(set) var items: [AddressBarPositionDisplayModel] = []
-
-    private let addressBarPositionManager: AddressBarPositionManaging
-
-    init(addressBarPositionManager: AddressBarPositionManaging = AppUserDefaults()) {
-        self.addressBarPositionManager = addressBarPositionManager
-        makeDisplayModels()
-    }
-
-    func setAddressBar(position: AddressBarPosition) {
-        addressBarPositionManager.currentAddressBarPosition = position
-        makeDisplayModels()
-    }
-
-    private func makeDisplayModels() {
-        items = AddressBarPosition.allCases.map { addressBarPosition in
-            let info = addressBarPosition.titleAndMessage
-
-            return AddressBarPositionDisplayModel(
-                type: addressBarPosition,
-                icon: addressBarPosition.image,
-                title: info.title,
-                message: info.message,
-                isSelected: addressBarPositionManager.currentAddressBarPosition == addressBarPosition
-            )
-        }
-    }
-}
-
-// MARK: - AddressBarPositionManaging
-
-protocol AddressBarPositionManaging: AnyObject {
-    var currentAddressBarPosition: AddressBarPosition { get set }
-}
-
-extension AppUserDefaults: AddressBarPositionManaging {}
-
-private extension AddressBarPosition {
-
-    var titleAndMessage: (title: NSAttributedString, message: String) {
-        switch self {
-        case .top:
-            (
-                NSAttributedString(string: UserText.HighlightsOnboardingExperiment.AddressBarPosition.topTitle),
-                UserText.HighlightsOnboardingExperiment.AddressBarPosition.topMessage
-            )
-        case .bottom:
-            (
-                NSAttributedString(string: UserText.HighlightsOnboardingExperiment.AddressBarPosition.bottomTitle),
-                UserText.HighlightsOnboardingExperiment.AddressBarPosition.bottomMessage
-            )
-        }
-    }
-
-    var image: ImageResource {
-        switch self {
-        case .top: .addressBarTop
-        case .bottom: .addressBarBottom
         }
     }
